+++ src/Media.php
@@ -182,6 +182,22 @@ class Media extends Model
     }
 
     /**
+     * Determine if the media item has a custom property with the given name
+     * using dot notation.
+     *
+     * @param string $propertyName
+     *
+     * @return bool
+     *
+     * @deprecated Will be removed in the next major version in favor of
+     * changing `hasCustomProperty` to use dot notation.
+     */
+    public function hasNestedCustomProperty(string $propertyName) : bool
+    {
+        return array_has($this->custom_properties, $propertyName);
+    }
+
+    /**
      * Get if the value of custom property with the given name.
      *
      * @param string $propertyName
@@ -195,6 +211,22 @@ class Media extends Model
     }
 
     /**
+     * Get a custom property using dot notation.
+     *
+     * @param string $propertyName
+     * @param mixed $default
+     *
+     * @return mixed
+     *
+     * @deprecated Will be removed in the next major version in favor of
+     * changing `getCustomProperty` to use dot notation.
+     */
+    public function getNestedCustomProperty(string $propertyName, $default = null)
+    {
+        return array_get($this->custom_properties, $propertyName, $default);
+    }
+
+    /**
      * @param string $name
      * @param mixed $value
      */
@@ -204,11 +236,31 @@ class Media extends Model
     }
 
     /**
+     * Set a custom property using dot notation.
+     *
+     * @param string $name
+     * @param mixed $value
+     *
+     * @deprecated Will be removed in the next major version in favor of
+     * changing `setCustomProperty` to use dot notation.
+     */
+    public function setNestedCustomProperty(string $name, $value)
+    {
+        // We need to assign `custom_properties` to a variable so we can
+        // modify it by reference.
+        $customProperties = $this->custom_properties;
+
+        array_set($customProperties, $name, $value);
+
+        $this->custom_properties = $customProperties;
+    }
+
+    /**
      * @param string $name
      *
      * @return $this
      */
-    public function removeCustomProperty(string $name)
+    public function forgetCustomProperty(string $name)
     {
         if ($this->hasCustomProperty($name)) {
             $customProperties = $this->custom_properties;
@@ -221,10 +273,42 @@ class Media extends Model
         return $this;
     }
 
+    /**
+     * @param string $name
+     *
+     * @return $this
+     *
+     * @deprecated Will be renamed to `forgetCustomProperty` in the next
+     * major version.
+     */
+    public function removeCustomProperty(string $name)
+    {
+        return $this->forgetCustomProperty($name);
+    }
+
+    /**
+     * Forget a custom property using dot notation.
+     *
+     * @param string $name
+     *
+     * @deprecated Will be removed in the next major version in favor of
+     * changing `forgetCustomProperty` to use dot notation.
+     */
+    public function forgetNestedCustomProperty(string $name)
+    {
+        // We need to assign `custom_properties` to a variable so we can
+        // modify it by reference.
+        $customProperties = $this->custom_properties;
+
+        array_forget($customProperties, $name);
+
+        $this->custom_properties = $customProperties;
+    }
+
     /*
      * Get all the names of the registered media conversions.
      */
-    public function getMediaConversionNames() : array
+    public function getMediaConversionNames(): array
     {
         $conversions = ConversionCollection::createForMedia($this);