+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to `laravel-medialibrary` will be documented in this file
 
-## 4.10.3 - 2017-01-14
+## 4.11.1 - 2017-01-14
 - fix bug in `setNewOrder`
+
+## 4.11.0 - 2017-01-10
+- added `hasNestedCustomProperty`, `getNestedCustomProperty`, `setNestedCustomProperty` and `forgetNestedCustomProperty` to use dot notation with custom properties
+- renamed `removeCustomProperty` to `forgetCustomProperty` (`removeCustomProperty` still exists but is marked as deprecated)
+
+## 4.10.3 - 2017-01-09
+- fix for getting preloaded media in the wrong order
 
 ## 4.10.2 - 2016-12-17
 - refactored the preloading of media